Factor away AbstractCloud
